requirements.txt
Django==1.8.6
django-taggit==0.17.1
pkg-resources==0.0.0
pytz==2018.7
